photo_backuper_app/requirements.txt
flask==3.0.0
flask_wtf==1.2.1
wtforms==3.1.2
python-dotenv=1.0.1